Vo相关部分修改完成 接下来就是设计钱包系统

+++ elmboot_SpringBoot/src/main/java/com/neusoft/elmboot/controller/DeliveryAddressController.java
@@ -2,6 +2,9 @@ package com.neusoft.elmboot.controller;
 
 import java.util.List;
 
+import com.neusoft.elmboot.exception.BusinessException;
+import com.neusoft.elmboot.exception.ErrorCode;
+import com.neusoft.elmboot.model.vo.DeliveryAddressVo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -15,30 +18,44 @@ public class DeliveryAddressController {
     private DeliveryAddressService deliveryAddressService;
 
     @RequestMapping("/listDeliveryAddressByUserId")
-    public List<DeliveryAddress> listDeliveryAddressByUserId(DeliveryAddress deliveryAddress)
+    public List<DeliveryAddressVo> listDeliveryAddressByUserId(DeliveryAddress deliveryAddress)
             throws Exception {
+        if (deliveryAddress.getUserId() == null) {
+            throw new BusinessException(ErrorCode.SYSTEM_ERROR);
+        }
         return deliveryAddressService.listDeliveryAddressByUserId(deliveryAddress.getUserId());
     }
 
     @RequestMapping("/getDeliveryAddressById")
-    public DeliveryAddress getDeliveryAddressById(DeliveryAddress deliveryAddress) throws
+    public DeliveryAddressVo getDeliveryAddressById(DeliveryAddress deliveryAddress) throws
             Exception {
+        if (deliveryAddress.getDaId() == null) {
+            throw new BusinessException(ErrorCode.SYSTEM_ERROR);
+        }
         return deliveryAddressService.getDeliveryAddressById(deliveryAddress.getDaId());
     }
 
     @RequestMapping("/saveDeliveryAddress")
     public int saveDeliveryAddress(DeliveryAddress deliveryAddress) throws Exception {
+        if (deliveryAddress.getContactName() == null || deliveryAddress.getContactSex() == null || deliveryAddress.getContactTel() == null || deliveryAddress.getAddress() == null) {
+            throw new BusinessException(ErrorCode.SYSTEM_ERROR);
+        }
         return deliveryAddressService.saveDeliveryAddress(deliveryAddress);
     }
 
     @RequestMapping("/updateDeliveryAddress")
     public int updateDeliveryAddress(DeliveryAddress deliveryAddress) throws Exception {
+        if (deliveryAddress.getContactName() == null || deliveryAddress.getContactSex() == null || deliveryAddress.getContactTel() == null || deliveryAddress.getAddress() == null) {
+            throw new BusinessException(ErrorCode.SYSTEM_ERROR);
+        }
         return deliveryAddressService.updateDeliveryAddress(deliveryAddress);
     }
 
     @RequestMapping("/removeDeliveryAddress")
     public int removeDeliveryAddress(DeliveryAddress deliveryAddress) throws Exception {
+        if (deliveryAddress.getDaId() == null) {
+            throw new BusinessException(ErrorCode.SYSTEM_ERROR);
+        }
         return deliveryAddressService.removeDeliveryAddress(deliveryAddress.getDaId());
-
     }
 }
